Revert "Get rid of bad lock"
This reverts commit 9e48bff7ad0c02043a7bbd035baef20ff29411fe.

Revert "cleanup"

This reverts commit e67cced581a8941d914af93d3bd4a687cf80251f.

Revert "Fix race condition"

This reverts commit 3f87bec3a62ffadd39741770386de6610451d314.

Revert "Missing test"

This reverts commit 16f1285555c482e1db37a0f8d6e1eb7072bd8cdb.

Revert "Break up tests into their categories instead of putting them all in one L1 file"

This reverts commit 245ded3c0b10cd64792bd6bddb7d02471dd62159.
